requirements.txt
Flask
flask-socketio
eventlet
pyserial
pandas
plotly
